web_app/requirements.txt
Flask==0.12
google-cloud-storage==1.42.0